update project scope
- include socom 2, 3 & CA references
- reorganize SOCOM header
- basic engine header for vectors and basic structures that are in all games
