+++ Main.cpp
@@ -1,5 +1,4 @@
 #include "Menu.h"
-#include "SOCOM.h"
 static bool bRunning{ true };
 
 int GameUpdate();
