+++ SOCOM-ESP.vcxproj
@@ -78,9 +78,11 @@
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
     <LinkIncremental>true</LinkIncremental>
+    <OutDir>$(SolutionDir)\bin\$(Platform)\$(Configuration)\</OutDir>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
     <LinkIncremental>false</LinkIncremental>
+    <OutDir>$(SolutionDir)\bin\$(Platform)\$(Configuration)\</OutDir>
   </PropertyGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
     <ClCompile>
@@ -154,12 +156,12 @@
     <ClCompile Include="libs\Dear ImGui\1.91.4\imgui_widgets.cpp" />
     <ClCompile Include="libs\DXWindow\DXWindow .cpp" />
     <ClCompile Include="libs\Memory\Memory.cpp" />
+    <ClCompile Include="libs\SDK\engine.cpp" />
+    <ClCompile Include="libs\SDK\SOCOM.cpp" />
     <ClCompile Include="Main.cpp" />
     <ClCompile Include="Menu.cpp" />
-    <ClCompile Include="SOCOM.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="SOCOM.h" />
     <ClInclude Include="Menu.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
